wo#6269 . split up kernel.c so that init_kernel() and references to kernel types is in a single file
